rename logger.js to raygun.client.js
