+++ examples/express-sample/app.js
@@ -4,7 +4,7 @@ var logger = require("morgan");
 var cookieParser = require("cookie-parser");
 var bodyParser = require("body-parser");
 var sassMiddleware = require("node-sass-middleware");
-var raygunClient = require("./logger");
+var raygunClient = require("./raygun.client");
 
 var routes = require("./routes/index");
 var users = require("./routes/users");
