+++ examples/express-sample/raygun.client.js
